remove version check
Remove the version check, in favor of just checking wether running the action results in any changes or not.

if there are changes, create a PR, if not do not create a PR.

This helps solve some issues with how people format their versions, and moves the responsibility to the workflow that uses the action to determine wether it should run or not.